Add files via upload
You are required to create a program which accepts DNA sequences as input and stores them in a list. These sequences consist of only ['A/a', 'C/c', 'G/g', 'T/t'] characters, but there is no guarantee that the user will not make mistakes. Once a sequence has been entered your program should first do a quality control step to ensure that the sequence is valid (only contains [A/a, C/c, G/g, T/t] characters). If the sequence is not valid display an error message. If the sequence is valid reverse complement it (Reverse the sequence and swap ‘A/a’ characters with ‘T/t’, ‘G/g’ characters with ‘C/c’ and vice versa) If the resulting sequence is already in the list of sequences alert the user that this is the case otherwise store the sequence in the list. The program should loop until the user inputs ‘done’ after which your program should print the sequences which have been stored. 
Optional: Print some statistics regarding the sequences. 
eg: 
frequency of each of the valid